requirements.txt
# pep8
black
pylint
# spider
requests
scrapy
scrapy-redis
scrapy-splash
# parse
lxml
# auto
selenium
pyppeteer
pyautogui
# js
pyexecjs
# database
pymysql
pymongo
redis
# data
pandas
# excel
xlrd
xlwt
xlutils
openpyxl
# word
python-docx
# api
fastapi
# asgi
uvicorn
# image
pillow
pdf2image
# pdf
pypdf2
pymupdf
pdfplumber
# ocr
pytesseract
# plot
setuptools
matplotlib
# exe
pyinstaller
# pyqt
pyqt5==5.13.0
pyqt5-stubs==5.13.0.1
pyqt5-tools==5.13.0.1.5
# other
retry
retrying
loguru
sshtunnel
pyyaml
demjson
cpca==0.5.0
folium
fonttools
psutil
# sdk
aliyun-python-sdk-core
aliyun-python-sdk-alimt
baidu-aip
tencentcloud-sdk-python


# tesserocr
# frida==12.8.0
# frida-tools==5.3.0
# objection==1.8.4
